Reading files is now well documented.
I moved the logic of reading files to its own function that takes a
path, since the three read file tests only differ based on the str
passed to it.

The documentation was restructed due to this change, and how to read
any file is unbraided from reading an absolute URL.

Writing files is next.